+++ actix-web-lab/src/extract.rs
@@ -7,6 +7,7 @@ pub type SharedData<T> = actix_web::web::Data<T>;
 pub use crate::body_hash::BodyHash;
 #[allow(deprecated)]
 pub use crate::body_hmac::{BodyHmac, HmacConfig};
+pub use crate::body_limit::BodyLimit;
 pub use crate::json::{Json, DEFAULT_JSON_LIMIT};
 pub use crate::lazy_data::LazyData;
 pub use crate::local_data::LocalData;